data: Move selected change path query
Selected-change state now stores hunk payloads, snapshots, and kind markers
in one data module, with file-change payloads in a separate module.

The project still resolves the current selected path inside the state
store, even though that query has to inspect hunk and file-change caches.

This commit continues splitting selected-change persistence by moving
selected path resolution into data.selected_change.paths and updating
source callers.

The next commit will update tests and architecture guards for the new path
query owner.
