+++ src/git_stage_batch/commands/discard.py
@@ -61,11 +61,11 @@ from ..data.selected_change.loading import (
 )
 from ..data.selected_change.store import (
     SelectedChangeKind,
-    get_selected_change_file_path,
     read_selected_change_kind,
     restore_selected_change_state,
     snapshot_selected_change_state,
 )
+from ..data.selected_change.paths import get_selected_change_file_path
 from ..data.selected_change.clear_reasons import (
     refuse_bare_action_after_auto_advance_disabled,
     refuse_bare_action_after_file_list,
