+++ src/git_stage_batch/commands/reset.py
@@ -54,9 +54,9 @@ from ..data.file_review.state import (
 from ..data.selected_change.lifecycle import clear_selected_change_state_files
 from ..data.selected_change.store import (
     SelectedChangeKind,
-    get_selected_change_file_path,
     read_selected_change_kind,
 )
+from ..data.selected_change.paths import get_selected_change_file_path
 from ..data.selected_change.clear_reasons import (
     mark_selected_change_cleared_by_stale_batch_selection,
 )
